Login and Home Pages
The following contains the web application that has a login that directs the user to the appropriate page. The home pages are for the client, technician and call service agent. Note that you will have to enter your database details for PostgreSQL in the ConnectionProvider file, it has been highlighted with dashes.